src/hashtable/WordPattern_290/SolutionTest.java
package hashtable.WordPattern_290;

import org.junit.Test;

import static org.junit.Assert.*;

public class SolutionTest {
    @Test
    public void test() {
        Solution s = new Solution();
        assertTrue(s.wordPattern("abba", "cat dog dog cat"));
        assertFalse(s.wordPattern("abba", "cat dog dog fish"));
        assertFalse(s.wordPattern("aaaa", "cat dog dog cat"));
        assertFalse(s.wordPattern("abba", "dog dog dog dog"));
        assertTrue(s.wordPattern("abc", "b c a"));
        assertTrue(s.wordPattern("ccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccdd",
                "s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s s t t"));
    }
}
